% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PokeDataAG-package.R
\docType{package}
\name{PokeDataAG-package}
\alias{PokeDataAG}
\alias{PokeDataAG-package}
\title{PokeDataAG: Analysis Group Training Dataset}
\description{
Contains datasets for Analysis Group R training.
    Datasets contain characteristics of various types of Pokemon.
}
\keyword{internal}
